updatesearch

+++ src/course/dao/StudentDaoInter.java
@@ -10,10 +10,10 @@ public interface StudentDaoInter {
     
     public boolean addStudent(Student s)throws Exception;
     
-    public Student getStudentById(Long id)throws Exception;
+    public Student getStudentById(int id)throws Exception;
     
-    public boolean getUpdateStudent(Student s, Long id)throws Exception;
+    public boolean getUpdateStudent(Student s, int id)throws Exception;
     
-    public boolean deleteStudent(Long studentId)throws Exception;
+    public boolean deleteStudent(int studentId)throws Exception;
     
 }
