+++ src/course/service/StudentServiceInter.java
@@ -3,16 +3,18 @@ package course.service;
 import course.model.Student;
 import java.util.List;
 
+
+
 public interface StudentServiceInter {
 
-    public List<Student> getStudentList() throws Exception;
-
-    public boolean addStudent(Student s) throws Exception;
-
-    public Student getStudentById(Long id) throws Exception;
-
-    public boolean getUpdateStudent(Student s, Long id) throws Exception;
-
-    public boolean deleteStudent(Long studentId) throws Exception;
-
+public List <Student> getStudentList()throws Exception;
+    
+    public boolean addStudent(Student s)throws Exception;
+    
+    public Student getStudentById(int id)throws Exception;
+    
+    public boolean getUpdateStudent(Student s, int id)throws Exception;
+    
+    public boolean deleteStudent(int studentId)throws Exception;
+    
 }
